upgrade module to PHP 7, extend test cases substantially, fix memory leaks
This is a major overhaul of nsphp, which makes it possible to use PHP 7
(tested with PHP 7.3.5). PHP 5 reached it EOL by jan 2019. PHP 7 has
substantial changes e.g. in memory and thread management compared
to PHP 5, several interfaces and API calls have changed, such that
it is not possible to compile the new version of nsphp with still
PHP 5. In case, someone wants to use nsphp with PHP 5, please
checkout versions before this commit from the repository.